Initial Commit of Northwinds-JSE
This is the initial commit of the Project. I have started working on this because, regardless of what I try, the NetBeans RCP is frustrating me at every turn. I just needed to get back to some plain old Java for a little bit.

Also, I have heard that porting an already written J2SE project with GUI to the NetBeans Platform is a little tricky, but things tend to work a little better. We will see, I guess.