Working on API
I am working on the API for the database access. If I can get it working reliably, it could be used in the Northwind Traders RCP project for data access for the Basic Edition, and, possibly, for the Corporate and Enterprise Editions as well, since I am really trying to abstract the access to a level that any driver will be able to be provided. I will continue working on this, but have some things that I need to try that were provided to me by Jaroslav Tulach on the NetBeans Developer email list. Due to this, I am going to pause work on these APIs while I try to get that going.