Changed Main Interface
I completely revamped the interface to a single window, with multiple tabs for the overview tables, a `JXTreeTable`, which is going to be for displaying the "per mile" breakdowns, and a status bar.

Each "module" of the application will still offer various dialogs, as needed. and then we will need to work on the reporting features.

I am trying some tricks from the Todo-NB sample application from the May, 2006, edition of NetBeans Magazine, written by Fernando Lazano. I am not sure how this is going to work out, but what the heck, I will give it a go.