Working on the Static PoC
Right now, I am working on the static Proof of Concept. Once I have the entire GUI designed, I will start working on converting the static PoC to a dynamic PoC. Then, I will create the business logic behind the system.

I need to remember to modify the call to the `log.welcome` method to add the "modules" as I get them finished for the static PoC. Once the static PoC is complete, and I start working on the dynamic PoC, I will delete the "modules" in that call, then add them back as I complete the dynamic capabilities of the PoC. By doing this, it will give me a way of tracking my progress on these PoCs.