Static PoC Work Continues
Working on tying various parts of the interface to the others. Already have the Load Task Pane and Load Tracker tab connected. When the user selects the Load Tracker tab, the Loads Task Pane expands and all the rest, except for the Northwind Traders Task Pane, collapse, since those tasks are not applicable to the Load Tracker tab. Likewise, when a different tab is selected, and the user expands the Loads Task Pane, the rest of the task panes, except the Northwind task pane, collapse and the Load Tracker tab becomes active. I will get this working for all tabs and all task panes. The Northwind Traders Task Pane will always be expanded (unless the user collapses it) because those tasks are global to the application.

I also modified the right side of the window to have a `JTextPane` in the lower-right corner that is capable of displaying short HTML help tips for assisting the user in performing their tasks. This will always be visible and provide quick information. I will create a full online help system once I have the dynamic PoC working. Once the dynamic PoC is working, I will send a link to @jkovalsy to fork the repo, test it for me, and provide me with feedback. Once I have his feedback, I can fix or enhance anything he points out, then have a version 1.0 of the Basic Edition, once I connect the database behind it.

Speaking of the database, I have started creating it in SQLite, as an embedded database and have included the SQLite JDBC driver in the libraries of the application. So, now we have the database drivers for:

* HSQLDB
* Derby
* PostGRE SQL
* MySQL
* SQLite

I may add the JDBC drivers for Microsoft SQL, as well, but have not decided yet.

Anyway, work continues and progress is being made.