Closing Tabs Issue and Fix
Since I wiped out the `mainToolbar` in favor of the `JXTaskPaneController` and `JXTaskPane`, I ran into a problem with "closing" and "opening" the "windows", a.k.a., removing and adding the `JPanel`s from and to the `JTabbedPane`. It seemed that whenever I closed a tab, then reopened it, I could not expand its related `JXTaskPane` and reactivate that restored tab. However, this evening, I figured out why and how to fix it.

I had been operating the connection between the tabs and task panes based upon the `mainTabbedPane.getSelectedIndex()` and `mainTabbedPane.setSelectedIndex(x)` methods. However, once you remove a tab from the tabbed pane, all the remaining tabs shift to the left, if necessary. Therefore, when you add that removed tab back, it now has an index of `JTabbedPane.getTabCount() + 1`.

The fix is to ***not*** use the `*selectedIndex` methods, but to use the `*selectedComponent` methods. So, after my discovery, this is what the code became:

```java
Component oldTab = mainTabbedPane.getSelectedComponent();

try {	// To catch IllegalArgumentException in the event the tab is closed.
    if ( !loadsTaskPane.isCollapsed ) { // Task pane must be expanded. 
		If ( !mainTabbedPane.getTitleAt(
				mainTabbedPane.getSelectedIndex()).equalsIgnoreCase(
						"Load Tracker") ) {
			mainTabbedPane.setSelectedComponent(loadsTab);
		} else {
			mainTabbedPane.setSelectedComponent(oldTab);
		}
	}
} catch ( IllegalArgumentException e ) {
	loadsTaskPane.setCollapsed(true);
}

this.mainTabbedPaneStateChanged(null);
```

Now the only issue is that if I use the menu to view and hide the tab, and had already once removed the tab via the task pane, the task pane's close command changes to open, but then never changes back. Need to look into this a little more to see how we can make the task change.
Signed-off-by:Integ <Integ@DESKTOP-5S661S1>