Versioning is Holding When NOT Debugging
I now have the automatic versioning holding to the proper build number when not debugging/working within the IDE. To make this work, I had to move the version calculation into its own class and out of the static initializer for the `AppProperties` class. Because of being in the static initializer for the `AppProperties` class, as soon as that class was accessed, it would run and would always get the default property value of `true` for the debugging query, therefore would always run the calculation, regardless of whether or not the debugging command-line argument (`--debugging` or `-d`) parameter was present.

Now, when the parameter is present, the calculation is run, as it should. However, if the parameter is not present the calculations do not run. My last clean and build to create the `Northwind-JSE.jar` file in the `{$PROJECT}/dist` folder is showing The following version:

0.1.10 Build 1824292

It shows this version and build regardless of how many times I run the `jar` file. This is the way it should be working!