Expanding Task Opens Closed Tab
I have changed the code to have a task pane expansion open the associated tab if that tab is closed. I felt that this was a better plan than showing an error message in the status bar and forcing the user to manually open the tab via the View menu. Since everything else regarding tabs and related task panes is somewhat automated, I think this is a better solution for closed tabs, since its related task pane remains in the task pane container.