Concentrating on the GUI
At this point, I am just going to concentrate on the GUI of the application to make it flow with the work, as best as possible. Once the GUI is working properly, I will start putting the data behind it, in a static manner, for the Proof of Concept. 

Part of getting the GUI working properly will involve data validation.