Trying to Modify the GUI on the Fly
I am attempting to get the dialog boxes to be undecorated on Linux and decorated on Windows. When they are undecorated on Linux, the drop shadow is still around them, but on Windows, there is no drop shadow unless they are decorated. Undecorated on Linux looks really nice, sort of like a popup on a web page, but on Windows it does not look good, so I want the borders and title bars to be present. I'm loading the system on Linux now and will get this figured out.
Signed-off-by:Integ <Integ@DESKTOP-5S661S1>